internal/blobtest: move blob testutils
Move blob-related testutils into a separate blobtest package. Future work will
require a transitive dependency on the sstable package (via internal/manifest).
